chore: refactor platform abstract
