+++ src/server/services/bot/platforms/discord/index.ts
@@ -1,0 +1,2 @@
+export { Discord, discordDescriptor } from './bot';
+export { DiscordRestApi } from './restApi';
