+++ src/server/services/bot/platforms/lark/index.ts
@@ -1,0 +1,2 @@
+export { feishuDescriptor, Lark, larkDescriptor } from './bot';
+export { LarkRestApi } from './restApi';
