+++ src/server/services/bot/platforms/telegram/index.ts
@@ -1,0 +1,2 @@
+export { setTelegramWebhook, Telegram, telegramDescriptor } from './bot';
+export { TELEGRAM_API_BASE, TelegramRestApi } from './restApi';
